+++ frontend/src/pages/Messages/AddMessage.tsx
@@ -3,7 +3,20 @@ import { Button, Grid } from "@mui/material";
 import TextField from "@mui/material/TextField";
 import { Socket } from "socket.io-client";
 
-export const AddMessageForm: React.FC<{socket: Socket | null}> = ({ socket }) => {
+interface IAddMessageForm {
+	socket: Socket | null;
+	userId: number | null;
+}
+
+export const AddMessageForm: React.FC<IAddMessageForm> = ({ socket, userId }) => {
+
+	const [message, setMessage] = React.useState('');
+
+	const onSubmit = () => {
+		socket?.emit('sendMessage', {  message, userId, conversationId: 3 })
+		setMessage('');
+	}
+
 
 	return <Grid>
 			<TextField
@@ -12,6 +25,8 @@ export const AddMessageForm: React.FC<{socket: Socket | null}> = ({ socket }) =>
 				id="message"
 				label="Message"
 				autoComplete="Message"
+				value={message}
+				onChange={(e) => setMessage(e.currentTarget.value)}
 			/>
 		<Button onClick={() => null} variant="contained">Send</Button>
 	</Grid>