+++ frontend/src/pages/Messages/ConversationList.tsx
@@ -1,20 +1,23 @@
 import React from "react";
 import { ConversationDto } from "../../types/message/conversation.dto";
 import { Grid, Typography } from "@mui/material";
+import { Link } from "react-router-dom";
 
 
 interface IConversations {
 	conversations: ConversationDto[];
 }
-export const Conversations:React.FC<IConversations> = ({ conversations }) => {
+
+export const ConversationList: React.FC<IConversations> = ({ conversations }) => {
 	return (
 		<Grid>
 			{conversations.map((conversation) => {
-				return <Grid>
+				return <Link to={`/${conversation.id}`}>
 					<Typography>{conversation.id}</Typography>
-				</Grid>
+				</Link>;
 			})}
 		</Grid>
 	);
 };
 
+;